[#53196411] Now able to pull form values on click
By removing the function paramters and simply grabbing the form values
within the click function I was able to get current values at time of
click.  Next is to put a little logic around if the username and
password are blank because I do not want to call the REST api without
credentials.  Then call the API to get the person token, store it the
use it for all subsequent requests.  Any time the credentials change we
use the new token for all subsequent requests.