[#53196411] Moved past injecting Base64 service
I still am not completely understanding how the DI works and the JS
syntax is definitely not my favorite.  I need a better understanding of
the Angular order of operations and specifically how services and
modules are registered and then found and loaded.  Currently am stuck
wondering why the http is not available.