[Finishes: #53196411]
OK so the POC is done, now its time to refactor and do asom serious
clean up.  We need to move the repetitive code up to have a base
controller that calls the CORS proxy.  The php cors proxy needs to
become a class for reusability.  Is it worth having the php class accept
the REST syntax so that the JS controllers can use the $resource object
provided by angular ?

I also need to create a local data store for caching so that we can
start answering some of questions.  Need to map each of the objects into
related tables so we can apply some SQL.

It would be nice to have an actual UI.

We need a charting library to feed data to so we can get nice graphs.